requirements.txt
Flask
line-bot-sdk
neologdn
emoji
mecab-python3
unidic-lite
markovify
